+++ src/lib/metrics/utils.ts
@@ -21,8 +21,11 @@ import {
   OrderMetricData,
   OrderMetricId,
   OrderMetricType,
+  OrderSentEvent,
+  OrderSimulatedEvent,
   OrderStage,
   OrderTxnFailedEvent,
+  OrderTxnSubmittedEvent,
   PendingTxnErrorEvent,
   ShiftGmMetricData,
   SwapGmMetricData,
@@ -422,6 +425,81 @@ export function sendOrderSubmittedMetric(metricId: OrderMetricId) {
   });
 
   metrics.startTimer(metricId);
+}
+
+export function sendOrderSimulatedMetric(metricId: OrderMetricId) {
+  const metricData = metrics.getCachedMetricData<OrderMetricData>(metricId);
+
+  if (!metricData) {
+    metrics.pushError("Order metric data not found", "sendOrderSimulatedMetric");
+    return;
+  }
+
+  metrics.pushEvent<OrderSimulatedEvent>({
+    event: `${metricData.metricType}.simulated`,
+    isError: false,
+    time: metrics.getTime(metricId)!,
+    data: metricData,
+  });
+}
+
+export function sendOrderSimulationErrorMetric(metricId: OrderMetricId, error: Error) {
+  const metricData = metrics.getCachedMetricData<OrderMetricData>(metricId);
+
+  if (!metricData) {
+    metrics.pushError("Order metric data not found", "sendOrderSimulationErrorMetric");
+    return;
+  }
+
+  const errorData = prepareErrorMetricData(error);
+
+  metrics.pushEvent<OrderTxnFailedEvent>({
+    event: `${metricData.metricType}.${OrderStage.Failed}`,
+    isError: true,
+    data: {
+      errorContext: "simulation",
+      ...(errorData || {}),
+      ...metricData,
+    },
+  });
+}
+
+export function sendOrderTxnSubmittedMetric(metricId: OrderMetricId) {
+  const metricData = metrics.getCachedMetricData<OrderMetricData>(metricId);
+
+  if (!metricData) {
+    metrics.pushError("Order metric data not found", "sendOrderTxnSubmittedMetric");
+    return;
+  }
+
+  metrics.pushEvent<OrderTxnSubmittedEvent>({
+    event: `${metricData.metricType}.txnSubmitted`,
+    isError: false,
+    time: metrics.getTime(metricId)!,
+    data: metricData,
+  });
+}
+
+export function makeTxnSentMetricsHandler(metricId: OrderMetricId) {
+  return () => {
+    const metricData = metrics.getCachedMetricData<OrderMetricData>(metricId);
+
+    if (!metricData) {
+      metrics.pushError("Order metric data not found", "makeTxnSentMetricsHandler");
+      return;
+    }
+
+    metrics.startTimer(metricId);
+
+    metrics.pushEvent<OrderSentEvent>({
+      event: `${metricData.metricType}.sent`,
+      isError: false,
+      time: metrics.getTime(metricId)!,
+      data: metricData,
+    });
+
+    return Promise.resolve();
+  };
 }
 
 export function sendTxnValidationErrorMetric(metricId: OrderMetricId) {
